formatting CMakeLists.txt in glfw_oldgl

+++ src_cpp/cpp/dtri_v3.cpp
@@ -169,7 +169,7 @@ static bool FindRayTriangleMeshIntersectionClosestToPoint
   }
   
     // Find the point that is the closest to the target point
-  float minSquareDistance = 1.0e16;
+  double minSquareDistance = 1.0e16;
   for (unsigned int i = 0; i < intersectionPoints.size(); i++)
   {
     float currSquareDistance =
